decrease the memory requirement in 3-D coupling: step 2

+++ src/CoR/remap_weight_of_strategy_class.h
@@ -34,8 +34,8 @@ class Remap_weight_of_operator_instance_class
         friend class Remap_weight_of_operator_class;
 		Remap_weight_of_operator_class *remap_weight_of_operator;
         Remap_operator_basis *duplicated_remap_operator;
-        long remap_beg_iter;
-        long remap_end_iter;
+        int remap_beg_iter;
+        int remap_end_iter;
         
     public: 
         Remap_weight_of_operator_instance_class() { duplicated_remap_operator = NULL; }
@@ -125,7 +125,7 @@ class Remap_weight_of_strategy_class
         void do_remap(int, Remap_grid_data_class*, Remap_grid_data_class*);
         void add_remap_weight_of_operator_instance(Remap_weight_of_operator_instance_class *, Remap_grid_class *, Remap_grid_class *, Remap_operator_basis *, Remap_grid_class *, Remap_grid_class *);
         void calculate_src_decomp(Remap_grid_class*, Remap_grid_class*, long*, const long*);
-        Remap_grid_class **get_remap_related_grids(int&);
+		Remap_grid_class **get_remap_related_grids(int&);
         Remap_weight_of_strategy_class *generate_parallel_remap_weights(Remap_grid_class**, Remap_grid_class**, int **);
         const char *get_object_name() { return object_name; }
         void renew_object_name(const char*);
